Rename: friend api 구조 수정
