+++ src/pages/Layout/MainLayout/MemberSection/SelectFriendSection/SelectFriendHeader/index.tsx
@@ -3,8 +3,8 @@ import { BiCheckCircle, BiUserCheck } from 'react-icons/bi'
 import { Box, Text, useDisclosure } from '@chakra-ui/react'
 import { useMutation, useQueryClient } from '@tanstack/react-query'
 
-import { addFriends } from '@/api/services/friend'
-import { friendsQueries } from '@/api/services/friend/useFriends'
+import { addFriends } from '@/api/services/friend/friend.api'
+import { friendsQueries } from '@/api/services/friend/queries'
 import { AlertModal } from '@/components/Modal/AlertModal'
 import { PageLayout } from '@/components/PageLayout'
 import { useFriendStore } from '@/stores/friends'
